requirements.txt
click==8.1.8
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.2
pyarrow==21.0.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
typer==0.16.1
tzdata==2025.2
